+++ packages/generator/__tests__/generate-token-js.test.ts
@@ -6,6 +6,10 @@ test('[dts] should generate package', () => {
   expect(generateTokenJs(generator).js).toMatchInlineSnapshot(
     `
     "const tokens = {
+      \\"borders.none\\": {
+        \\"value\\": \\"none\\",
+        \\"variable\\": \\"var(--borders-none)\\"
+      },
       \\"easings.default\\": {
         \\"value\\": \\"cubic-bezier(0.4, 0, 0.2, 1)\\",
         \\"variable\\": \\"var(--easings-default)\\"
@@ -962,6 +966,10 @@ test('[dts] should generate package', () => {
         \\"value\\": \\"64px\\",
         \\"variable\\": \\"var(--blurs-3xl)\\"
       },
+      \\"spacing.0\\": {
+        \\"value\\": \\"0rem\\",
+        \\"variable\\": \\"var(--spacing-0)\\"
+      },
       \\"spacing.1\\": {
         \\"value\\": \\"0.25rem\\",
         \\"variable\\": \\"var(--spacing-1)\\"
@@ -1093,6 +1101,10 @@ test('[dts] should generate package', () => {
       \\"spacing.3.5\\": {
         \\"value\\": \\"0.875rem\\",
         \\"variable\\": \\"var(--spacing-3\\\\\\\\.5)\\"
+      },
+      \\"sizes.0\\": {
+        \\"value\\": \\"0rem\\",
+        \\"variable\\": \\"var(--sizes-0)\\"
       },
       \\"sizes.1\\": {
         \\"value\\": \\"0.25rem\\",