+++ packages/preset-panda/src/borders.ts
@@ -1,0 +1,3 @@
+export const borders = {
+  none: { value: 'none' },
+}