+++ packages/studio/styled-system/css/index.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 export * from './css'
 export * from './cx'
 export * from './cva'