+++ packages/studio/styled-system/jsx/float.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import type { FunctionComponent } from 'react'
 import type { FloatProperties } from '../patterns/float'
 import type { HTMLPandaProps } from '../types/jsx'