+++ packages/studio/styled-system/jsx/spacer.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import type { FunctionComponent } from 'react'
 import type { SpacerProperties } from '../patterns/spacer'
 import type { HTMLPandaProps } from '../types/jsx'