+++ packages/studio/styled-system/patterns/aspect-ratio.mjs
@@ -2,17 +2,27 @@ import { mapObject } from '../helpers.mjs'
 import { css } from '../css/index.mjs'
 
 const aspectRatioConfig = {
-  transform(props) {
-    const { ratio, ...rest } = props
+  transform(props, { map }) {
+    const { ratio = 4 / 3, ...rest } = props
     return {
-      aspectRatio: ratio,
-      overflow: 'hidden',
-      display: 'flex',
-      justifyContent: 'center',
-      alignItems: 'center',
-      '&>img, &>video': {
+      position: 'relative',
+      _before: {
+        content: `""`,
+        display: 'block',
+        height: '0',
+        paddingBottom: map(ratio, (r) => `${(1 / r) * 100}%`),
+      },
+      '&>*': {
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        overflow: 'hidden',
+        position: 'absolute',
+        inset: '0',
         width: '100%',
         height: '100%',
+      },
+      '&>img, &>video': {
         objectFit: 'cover',
       },
       ...rest,