+++ packages/studio/styled-system/tokens/index.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import type { Token } from './tokens'
 
 export declare const token: {