+++ packages/studio/styled-system/types/composition.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import type { CompositionStyleObject } from './system-types'
 
 type Recursive<T> = {