+++ packages/studio/styled-system/types/selectors.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import type { Pseudos } from './csstype'
 
 type AriaAttributes =
@@ -52,5 +53,7 @@ type DataAttributes =
 type AttributeSelector = `&${Pseudos | DataAttributes | AriaAttributes}`
 type ParentSelector = `${DataAttributes | AriaAttributes} &`
 
-export type AnySelector = `${string}&` | `&${string}`
+type AtRuleType = 'media' | 'layer' | 'container' | 'supports' | 'page'
+
+export type AnySelector = `${string}&` | `&${string}` | `@${AtRuleType}${string}`
 export type Selectors = AttributeSelector | ParentSelector