+++ CHANGELOG.md
@@ -1,8 +1,47 @@
 # CHANGELOG
 
+All notable changes to this project will be documented in this file.
+
 ## [Unreleased]
 
-- Nothing yet 😀
+See the [Changesets](./.changeset) for the latest changes.
+
+## [0.4.0]
+
+### Fixed
+
+- Fix issue here divider pattern generated incorrect css in horizontal orientation
+- Fix issue where aspect ratio css property adds `px`
+- Fix placeholder condition to map to `&::placeholder`
+- Fix issue where patterns that include css selectors doesn't work in JSX
+- Fix issue where the `panda ship` command does not write to the correct path
+
+### Added
+
+- Experimental support for native `.vue` files and `.svelte` files
+- Add types for supported at-rules (`@media`, `@layer`, `@container`, `@supports`, and `@page`)
+- Add webkit polyfill for common properties to reduce the need for `autoprefixer`
+- Add support for watch mode in codegen command via the `--watch` or `-w` flag.
+
+  ```sh
+  panda codegen --watch
+  ```
+
+- Add support for disabling shorthand props
+
+  ```ts
+  import { defineConfig } from '@pandacss/dev'
+
+  export default defineConfig({
+    // ...
+    shorthands: false,
+  })
+  ```
+
+### Changed
+
+- Add `auto` value where neccessary to base utilities.
+- Add `0` value to default spacing tokens to allow for `strictTokens` mode.
 
 ## [0.3.2]