+++ packages/extractor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pandacss/extractor
 
+## 0.4.0
+
+### Patch Changes
+
+- @pandacss/logger@0.4.0
+
 ## 0.3.2
 
 ### Patch Changes