+++ packages/postcss/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pandacss/postcss
 
+## 0.4.0
+
+### Patch Changes
+
+- @pandacss/node@0.4.0
+
 ## 0.3.2
 
 ### Patch Changes