+++ packages/preset-base/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pandacss/preset-base
 
+## 0.4.0
+
+### Patch Changes
+
+- e8024347: Fix issue here divider pattern generated incorrect css in horizontal orientation
+- d00eb17c: Add `auto` value where neccessary to base utilities.
+- 9156c1c6: Fix placeholder condition to map to `&::placeholder`
+- 54a8913c: Fix issue where patterns that include css selectors doesn't work in JSX
+- 0f36ebad: Add polyfill for common properties to reduce the need for autoprefixer
+- Updated dependencies [c7b42325]
+- Updated dependencies [5b344b9c]
+  - @pandacss/types@0.4.0
+
 ## 0.3.2
 
 ### Patch Changes